backend/requirements.txt
flask==2.3.2
ultralytics>=8.0.0
torchvision>=0.8.0
numpy>=1.18.5
torch>=2.0.0
transformers>=4.37.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
Pillow>=10.0.0
opencv-python>=4.8.0
argparse